requirements.txt
flask==3.0.0
line-bot-sdk==3.5.1
pdf2image==1.17.0
Pillow==10.2.0
requests==2.31.0
gunicorn==21.2.0
